requirements.txt
aiohttp==3.9.1
motor==3.3.2
pymongo==4.6.1
fastapi==0.109.2
uvicorn==0.27.1
feedparser==6.0.11
python-dateutil==2.8.2
python-dotenv==1.0.0
beautifulsoup4==4.12.3
pytz==2024.1 